refactored server.js to account for changes in model
